requirements.txt
Flask
pdfplumber
python-docx
pytesseract
Pillow
transformers
torch
spacy
pandas
mysql-connector-python